# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Vulnerability Scanner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ScanInput {
  /** The target URL to scan (must be a valid http/https URL) */
  url: string;
  /** When true, the scanner first spiders the target domain to discover internal links, forms, and endpoints, then runs injection checks against all discovered URLs with query parameters. */
  crawl_enabled?: boolean;
}

export type ScanFindingCategory = typeof ScanFindingCategory[keyof typeof ScanFindingCategory];


export const ScanFindingCategory = {
  dns: 'dns',
  ssl: 'ssl',
  headers: 'headers',
  server_info: 'server_info',
  content_discovery: 'content_discovery',
  http_methods: 'http_methods',
  sensitive_files: 'sensitive_files',
  injection: 'injection',
} as const;

export type ScanFindingSeverity = typeof ScanFindingSeverity[keyof typeof ScanFindingSeverity];


export const ScanFindingSeverity = {
  info: 'info',
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type ScanFindingStatus = typeof ScanFindingStatus[keyof typeof ScanFindingStatus];


export const ScanFindingStatus = {
  pass: 'pass',
  fail: 'fail',
  warning: 'warning',
  info: 'info',
} as const;

export interface ScanFinding {
  id: string;
  category: ScanFindingCategory;
  title: string;
  severity: ScanFindingSeverity;
  status: ScanFindingStatus;
  description: string;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  recommendation?: string | null;
  /**
     * Step-by-step proof-of-concept: how an attacker or tester would verify and exploit this finding, including commands (curl, Burp Suite, browser) and expected outcome.
     * @nullable
     */
  execution_poc?: string | null;
}

export type ScanResultSummary = {
  total: number;
  critical: number;
  high: number;
  medium: number;
  low: number;
  info: number;
  passed: number;
};

export type ScanResultCrawlSummary = {
  pagesVisited?: number;
  urlsDiscovered?: number;
  urlsWithParams?: string[];
  urls?: string[];
  jsFiles?: string[];
  errors?: string[];
} | null;

export interface ScanResult {
  findings: ScanFinding[];
  summary: ScanResultSummary;
  responseTimeMs: number;
  /** @nullable */
  serverInfo?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  sslValid?: boolean | null;
  /** @nullable */
  sslExpiry?: string | null;
  scannedAt: string;
  crawlSummary?: ScanResultCrawlSummary;
}

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Scan {
  id: number;
  url: string;
  status: ScanStatus;
  result?: ScanResult;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface ScanStats {
  totalScans: number;
  completedScans: number;
  failedScans: number;
  criticalFindings: number;
  highFindings: number;
  mediumFindings: number;
  lowFindings: number;
  recentScans: Scan[];
}