# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Web Vulnerability Scanner API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scans
    description: Vulnerability scan operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List scan history
      description: Returns all past scans ordered by newest first
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
    post:
      operationId: createScan
      tags: [scans]
      summary: Start a new scan
      description: Triggers a full vulnerability scan for the given URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Scan started/completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScan
      tags: [scans]
      summary: Delete a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully

  /scans/stats:
    get:
      operationId: getScanStats
      tags: [scans]
      summary: Aggregated scan statistics
      description: Returns counts by severity, total scans, and recent activity
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ScanInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: The target URL to scan (must be a valid http/https URL)
        crawl_enabled:
          type: boolean
          description: "When true, the scanner first spiders the target domain to discover internal links, forms, and endpoints, then runs injection checks against all discovered URLs with query parameters."

    ScanFinding:
      type: object
      required:
        - id
        - category
        - title
        - severity
        - status
        - description
      properties:
        id:
          type: string
        category:
          type: string
          enum: [dns, ssl, headers, server_info, content_discovery, http_methods, sensitive_files, injection]
        title:
          type: string
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        status:
          type: string
          enum: [pass, fail, warning, info]
        description:
          type: string
        detail:
          type: ["string", "null"]
        recommendation:
          type: ["string", "null"]
        execution_poc:
          type: ["string", "null"]
          description: "Step-by-step proof-of-concept: how an attacker or tester would verify and exploit this finding, including commands (curl, Burp Suite, browser) and expected outcome."

    ScanResult:
      type: object
      required:
        - findings
        - summary
        - responseTimeMs
        - scannedAt
      properties:
        findings:
          type: array
          items:
            $ref: "#/components/schemas/ScanFinding"
        summary:
          type: object
          required:
            - total
            - critical
            - high
            - medium
            - low
            - info
            - passed
          properties:
            total:
              type: integer
            critical:
              type: integer
            high:
              type: integer
            medium:
              type: integer
            low:
              type: integer
            info:
              type: integer
            passed:
              type: integer
        responseTimeMs:
          type: integer
        serverInfo:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        sslValid:
          type: ["boolean", "null"]
        sslExpiry:
          type: ["string", "null"]
        scannedAt:
          type: string
        crawlSummary:
          type: object
          nullable: true
          properties:
            pagesVisited:
              type: integer
            urlsDiscovered:
              type: integer
            urlsWithParams:
              type: array
              items:
                type: string
            urls:
              type: array
              items:
                type: string
            jsFiles:
              type: array
              items:
                type: string
            errors:
              type: array
              items:
                type: string

    Scan:
      type: object
      required:
        - id
        - url
        - status
        - createdAt
      properties:
        id:
          type: integer
        url:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        result:
          $ref: "#/components/schemas/ScanResult"
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]

    ScanStats:
      type: object
      required:
        - totalScans
        - completedScans
        - failedScans
        - criticalFindings
        - highFindings
        - mediumFindings
        - lowFindings
        - recentScans
      properties:
        totalScans:
          type: integer
        completedScans:
          type: integer
        failedScans:
          type: integer
        criticalFindings:
          type: integer
        highFindings:
          type: integer
        mediumFindings:
          type: integer
        lowFindings:
          type: integer
        recentScans:
          type: array
          items:
            $ref: "#/components/schemas/Scan"