# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Vulnerability Scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanInput {
  /** The target URL to scan (must be a valid http/https URL) */
  url: string;
  /** When true, the scanner first spiders the target domain to discover internal links, forms, and endpoints, then runs injection checks against all discovered URLs with query parameters. */
  crawl_enabled?: boolean;
}