# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Vulnerability Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanFinding } from './scanFinding';
import type { ScanResultCrawlSummary } from './scanResultCrawlSummary';
import type { ScanResultSummary } from './scanResultSummary';

export interface ScanResult {
  findings: ScanFinding[];
  summary: ScanResultSummary;
  responseTimeMs: number;
  /** @nullable */
  serverInfo?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  sslValid?: boolean | null;
  /** @nullable */
  sslExpiry?: string | null;
  scannedAt: string;
  crawlSummary?: ScanResultCrawlSummary;
}